test: factor out globals tests
